requirements.txt
quart
loguru
openai
pyserial